move get function to static.
